fix feature break from October VSCode insider
